+++ lib/src/socket/ssh_socket.dart
@@ -1,8 +1,8 @@
 import 'dart:async';
 import 'dart:typed_data';
 
-import 'package:dartssh2/src/socket/ssh_socket_io.dart'
-    if (dart.library.js) 'package:dartssh2/src/socket/ssh_socket_js.dart';
+import 'package:dartssh2/src/socket/ssh_socket_js.dart'
+    if (dart.library.io) 'package:dartssh2/src/socket/ssh_socket_io.dart';
 
 abstract class SSHSocket {
   /// Connects using the platform native socket transport.